Add: User Auth endpoint and relationships

+++ src/controllers/servicesController.ts
@@ -5,6 +5,7 @@ import Service from '../models/Service';
 import slugify from 'slugify';
 import AppError from '../utils/AppError';
 import Entry from '../models/Entry';
+import User from '../models/User';
 
 export const getServices = catchAsync(async (req: Request, res: Response, next: NextFunction) => {
   const services = await Service.find().populate('entries');
@@ -25,7 +26,17 @@ export const getService = catchAsync(async (req: Request, res: Response, next: N
 export const createService = catchAsync(async (req: Request, res: Response, next: NextFunction) => {
   const request = createServiceSchema.parse(req.body);
 
-  const service = await Service.create({ ...request, user: req.user?._id });
+  const user = await User.findById(req.user?.id);
+
+  if (!user) {
+    return next(new AppError('User not found', 404));
+  }
+
+  const service = await Service.create({ ...request, user: user.id });
+
+  user.services.push(service);
+
+  await user.save();
 
   res.status(201).json({ status: 'success', service });
 });
@@ -57,9 +68,22 @@ export const deleteService = catchAsync(async (req: Request, res: Response, next
     return next(new AppError('Service not found', 404));
   }
 
-  await Entry.deleteMany({ service });
+  const user = await User.findById(req.user?.id);
 
-  await service.deleteOne();
+  if (!user) {
+    return next(new AppError('User not found', 404));
+  }
+
+  if (user.services.includes(service.id)) {
+    const serviceIdx = user.services.indexOf(service.id);
+
+    user.services.splice(serviceIdx, 1);
+
+    await user.save();
+  }
+
+  await Entry.deleteMany({ service });
+  await service?.deleteOne();
 
   res.status(200).json({ status: 'success', service: null });
 });
